タスクリストの列順を変更。列定義関連をclass分離。
