+++ ProjectsTM.UI.TaskList/ProjectsTM.UI.TaskList.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ColDefinition.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TaskListForm.cs">
       <SubType>Form</SubType>
